requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
apturl==0.5.2
asgiref==3.8.1
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
build==0.10.0
CacheControl==0.12.14
cachy==0.3.0
certifi==2024.2.2
chardet==4.0.0
charset-normalizer==3.1.0
cleo==0.8.1
click==8.1.7
clikit==0.6.2
colorama==0.4.4
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
crashtest==0.3.1
cryptography==3.4.8
cupshelpers==1.0
DateTime==4.7
dbus-python==1.2.18
defer==1.0.6
deta==1.2.0
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.3
django-debug-toolbar==4.3.0
django-filter==24.1
django-js-asset==2.2.0
django-mptt==0.16.0
django-ranged-response==0.2.0
django-rest-auth==0.9.5
django-simple-captcha==0.6.0
-e git+https://github.com/BCSDev8/simpleui.git@6804f669dbf21a8454ed1fd9dc0899f045a733bd#egg=django_simpleui
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docker==5.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.21.5
dulwich==0.21.6
duplicity==0.8.21
exceptiongroup==1.2.0
fastapi==0.109.2
fasteners==0.14.1
filelock==3.12.4
frozenlist==1.3.0
future==0.18.2
greenlet==3.0.3
gyp==0.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.4
httplib2==0.20.2
httpx==0.27.0
idna==3.6
importlib-metadata==6.0.0
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
installer==0.7.0
itypes==1.2.0
jaraco.classes==3.2.3
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==3.2.0
keyring==21.8.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.3.2
MarkupSafe==2.1.5
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.4
multidict==6.0.2
netifaces==0.11.0
numpy==1.23.1
oauthlib==3.2.0
olefile==0.46
opencv-python==4.6.0.66
packaging==23.2
paramiko==2.9.3
pastel==0.2.1
pexpect==4.8.0
pillow==10.2.0
pkginfo==1.9.6
platformdirs==3.11.0
pluggy==1.4.0
poetry==1.1.12
poetry-core==1.0.8
poetry-plugin-export==1.5.0
protobuf==3.12.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.6.2
pydantic_core==2.16.3
PyGObject==3.42.1
PyJWT==2.8.0
pylev==1.2.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyproject_hooks==1.0.0
pyRFC3339==1.1
pyrsistent==0.19.3
pytest==8.0.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==0.21.1
pytz==2023.3
pyxdg==0.27
PyYAML==5.4.1
rapidfuzz==2.13.7
reportlab==3.6.8
requests==2.31.0
requests-toolbelt==0.9.1
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
SecretStorage==3.3.1
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
source==1.2.0
SQLAlchemy==2.0.27
sqlparse==0.4.4
starlette==0.36.3
systemd-python==234
texttable==1.7.0
tomli==2.0.1
tomlkit==0.11.6
trove-classifiers==2023.1.12
typing_extensions==4.10.0
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.0.3
usb-creator==0.3.7
uvicorn==0.27.1
virtualenv==20.24.5
wadllib==1.3.6
webencodings==0.5.1
websocket-client==0.59.0
xdg==5
xkit==0.0.0
yarl==1.7.2
zipp==1.0.0
zope.interface==5.4.0
